ai-rag/requirements.txt
fastapi==0.104.1
uvicorn==0.24.0
pinecone-client==5.0.0
requests==2.31.0
pydantic==2.5.0

